make route for promo_code and add id check in redem button

+++ database/migrations/2025_02_12_070803_create_promo_code_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('promo_code', function (Blueprint $table) {
             $table->id();
             $table->string('code');
-            $table->decimal('discount_percentage')->nullable();
+            $table->integer('discount_percentage')->nullable();
             $table->timestamps();
         });
     }
